requirements.txt
asgiref
django >3.2,<3.3
pillow
pytz
sqlparse
typing_extensions
python-decouple
djangorestframework==3.14.0
django-filter 
drf-yasg[validation]
djangorestframework-simplejwt
django-mail-templated

#testing
pytest
pytest-django
pytest-cov
black
flake8